requirements.txt
#pip install -r requirements.txt
# add FOLDER: media, prewiev, image
asgiref==3.6.0
Django==4.1
Pillow==9.5.0
sqlparse==0.4.4
psycopg2-binary==2.9.5
gunicorn~=20.1.0